clean up StepEditor
