+++ javascript/react/src/components/StepEditor/DeleteButton.styles.ts
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+
+export const icons = {
+	delete: { iconName: 'Delete' },
+}
