+++ javascript/react/src/components/StepEditor/DeleteButton.types.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export interface DeleteButtonProps {
+	onClick?: () => void
+}
