+++ javascript/react/src/components/StepEditor/SaveButton.styles.ts
@@ -1,0 +1,16 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+
+import styled from '@essex/styled-components'
+
+export const icons = {
+	checkMark: { iconName: 'CheckMark' },
+}
+
+export const SaveButtonWrapper = styled.div`
+	display: flex;
+	align-items: center;
+	gap: 0.5rem;
+`
