+++ javascript/react/src/components/StepEditor/SaveButton.types.ts
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export interface SaveButtonProps {
+	onClick?: () => void
+	disabled?: boolean
+}
